failsafe to prevent multiple alarms
I just did a 24 hr test with this and it appears to work. The ID isn't
really necessary, could have left it as 0. This flag is used instead of
FLAG_CANCEL_CURRENT because stack overflow said so and the docs suggest
FLAG_UPDATE_CURRENT when you don't care about the extra features garbage